Move ws logic to an abstract class
